refactor: PrefectureCheckboxをcomponents/atomsに移動
